+++ geotrek/settings/base.py
@@ -422,8 +422,8 @@ ALTIMETRIC_AREA_MARGIN = 0.15
 LEAFLET_CONFIG = {
     'SRID': SRID,
     'TILES': [
-        ('Scan', 'http://{s}.tiles.openstreetmap.org/{z}/{x}/{y}.png',),
-        ('Ortho', 'http://{s}.tiles.openstreetmap.org/{z}/{x}/{y}.jpg'),
+        ('Scan', 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',),
+        ('Ortho', 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.jpg'),
     ],
     'TILES_EXTENT': SPATIAL_EXTENT,
     # Extent in API projection (Leaflet view default extent)
@@ -546,3 +546,11 @@ SPLIT_TREKS_CATEGORIES_BY_ACCESSIBILITY = False
 
 CRISPY_ALLOWED_TEMPLATE_PACKS = ('bootstrap', 'bootstrap3')
 CRISPY_TEMPLATE_PACK = 'bootstrap'
+
+# Mobile app_directories
+MOBILE_TILES_URL = 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
+MOBILE_TILES_RADIUS_LARGE = 0.01  # ~1 km
+MOBILE_TILES_RADIUS_SMALL = 0.005  # ~500 m
+MOBILE_TILES_GLOBAL_ZOOMS = range(13)
+MOBILE_TILES_LOW_ZOOMS = range(13, 15)
+MOBILE_TILES_HIGH_ZOOMS = range(15, 17)