+++ geotrek/trekking/management/commands/sync_rando.py
@@ -19,6 +19,7 @@ from geotrek.common import models as common_models
 from geotrek.trekking import models as trekking_models
 from geotrek.common.views import DocumentPublicPDF
 from geotrek.trekking.views import TrekViewSet, POIViewSet, TrekPOIViewSet, TrekGPXDetail, TrekKMLDetail
+from geotrek.tourism.views import TrekTouristicContentAndPOIViewSet
 from geotrek.flatpages.views import FlatPageViewSet
 
 # Register mapentity models
@@ -168,9 +169,16 @@ class Command(BaseCommand):
         self.sync_view(lang, view, name, url='/?format=geojson', zipfile=zipfile)
 
     def sync_trek_pois(self, lang, trek, zipfile=None):
-        view = TrekPOIViewSet.as_view({'get': 'list'})
-        name = os.path.join('api', lang, 'treks', str(trek.pk), 'pois.geojson')
-        self.sync_view(lang, view, name, url='/?format=geojson', zipfile=zipfile, pk=trek.pk)
+        if settings.ZIP_TOURISTIC_CONTENTS_AS_POI:
+            view = TrekTouristicContentAndPOIViewSet.as_view({'get': 'list'})
+            name = os.path.join('api', lang, 'treks', str(trek.pk), 'pois.geojson')
+            self.sync_view(lang, view, name, url='/?format=geojson', zipfile=zipfile, pk=trek.pk)
+            view = TrekPOIViewSet.as_view({'get': 'list'})
+            self.sync_view(lang, view, name, url='/?format=geojson', zipfile=None, pk=trek.pk)
+        else:
+            view = TrekPOIViewSet.as_view({'get': 'list'})
+            name = os.path.join('api', lang, 'treks', str(trek.pk), 'pois.geojson')
+            self.sync_view(lang, view, name, url='/?format=geojson', zipfile=zipfile, pk=trek.pk)
 
     def sync_object_view(self, lang, obj, view, basename_fmt, zipfile=None, **kwargs):
         modelname = obj._meta.model_name