+++ docs/changelog.rst
@@ -13,6 +13,7 @@ CHANGELOG
 
 * Import framework
 * Static API to disconnect Geotrek-rando from Geotrek-Admin (fixes #1428)
+* Build zip files for mobile application
 
 **Bug fixes**