+++ docs/changelog.rst
@@ -17,6 +17,8 @@ CHANGELOG
 * Trek / Touristic content association distance depending on trek practice
 * Option to hide published treks nearby topologies
 * Add previous/next treks in API
+* Add external id in trekking/tourism detail pages and exports
+* Zip touristic contents as POI for mobile app v1
 
 **Bug fixes**